+++ dist_node/spec/cont.js
@@ -1,0 +1,12 @@
+/*
+ * THIS FILE IS AUTO GENERATED FROM 'lib/spec/cont.kep'
+ * DO NOT EDIT
+*/
+"use strict";
+var ContMonad;
+(ContMonad = (function(instance, callcc) {
+    (instance.prototype.callcc = callcc);
+    (instance.callcc = instance.prototype.callcc);
+    return instance;
+}));
+(module.exports = ContMonad);