+++ dist_node/spec/either.js
@@ -1,0 +1,13 @@
+/*
+ * THIS FILE IS AUTO GENERATED FROM 'lib/spec/either.kep'
+ * DO NOT EDIT
+*/
+"use strict";
+var EitherMonad;
+(EitherMonad = (function(Instance, left, right) {
+    (Instance.prototype.right = right);
+    (Instance.right = Instance.prototype.right);
+    (Instance.prototype.left = left);
+    (Instance.left = Instance.prototype.left);
+}));
+(module.exports = EitherMonad);