+++ dist_node/state.js
@@ -3,22 +3,22 @@
  * DO NOT EDIT
 */
 "use strict";
-var Trampoline = require("./trampoline"),
+var Identity = require("./identity"),
     StateT = require("./trans/state"),
     State;
-(State = StateT(Trampoline));
+(State = StateT(Identity));
 var x = StateT.runStateT,
-    y = Trampoline.run;
+    y = Identity.runIdentity;
 (State.runState = (function() {
     return y(x.apply(null, arguments));
 }));
 var x0 = StateT.evalStateT,
-    y0 = Trampoline.run;
+    y0 = Identity.runIdentity;
 (State.evalState = (function() {
     return y0(x0.apply(null, arguments));
 }));
 var x1 = StateT.execStateT,
-    y1 = Trampoline.run;
+    y1 = Identity.runIdentity;
 (State.execState = (function() {
     return y1(x1.apply(null, arguments));
 }));