+++ dist_node/trans/error.js
@@ -8,17 +8,13 @@ var __o = require("../structure"),
     Monoid = __o["Monoid"],
     EitherT = require("./either"),
     eitherT = EitherT["eitherT"],
+    ErrorMonad = require("../spec/error"),
     ErrorT;
 (ErrorT = (function(m) {
     var Instance = EitherT(m);
-    (Instance.prototype.fail = Instance.left);
-    (Instance.fail = Instance.prototype.fail);
-    (Instance.handle = (function(m0, e) {
-        return ErrorT.runErrorT(m0, m0.of, e);
-    }));
-    (Instance.prototype.handle = (function(e) {
-        var m0 = this;
-        return Instance.handle(m0, e);
+    ErrorMonad(Instance, Instance.left, (function(e) {
+        var c = this;
+        return ErrorT.runErrorT(c, c.of, e);
     }));
     return Instance;
 }));